new test for translation pairs
old 'dictionary' are going to be discarded and
replaced with a feature which shows which part are
related between original and translated text.

creating SPANs for every word (or character in
translated) is getting more and more silly,
not if cumbersome, so I'll try to use nodes
only when neccessary.